+++ src/petro/extractor/aac_sample_extractor.hpp
@@ -33,7 +33,7 @@ struct aac_sample_extractor : extractor
     /// Open this an the underlying layers
     void open(
         const uint8_t* data,
-        uint32_t size,
+        uint64_t size,
         uint32_t track_id,
         std::error_code& error) override;
