+++ src/petro/extractor/avc_sample_extractor.cpp
@@ -40,7 +40,7 @@ avc_sample_extractor::~avc_sample_extractor()
 
 void avc_sample_extractor::open(
         const uint8_t* data,
-        uint32_t size,
+        uint64_t size,
         uint32_t track_id,
         std::error_code& error)
 {
